+++ snisku/__init__.py
@@ -26,8 +26,8 @@ from . import kvs
 from . import param
 from . import types
 from . import ui
-from . import whitelist
+from . import whitelist  # Deprecated.
 
 __all__: Sequence[str] = ("argparse", "exc", "kvs", "param", "types", "ui",
                           "whitelist")
-__version__ = '0.2.0'
+__version__ = '0.3.0-SNAPSHOT'